# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Svloon",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/quick-services/{id}": {
            "get": {
                "tags": [
                    "QuickService"
                ],
                "summary": "getQuickServiceItem",
                "description": "Get QuickService",
                "operationId": "d5a8ff13ffd2f974d4150d91b792698b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of QuickService",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/QuickService"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "QuickService"
                ],
                "summary": "updateQuickService",
                "description": "Update QuickService",
                "operationId": "3d6cb38671f6063cc772e0acffe3cd2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of QuickService",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuickService"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/QuickService"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "QuickService"
                ],
                "summary": "deleteQuickService",
                "description": "Delete QuickService",
                "operationId": "6a5a9732ec7f59faef8f327fe8bf9a23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of QuickService",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "QuickService": {
                "required": [],
                "properties": {
                    "service_id": {
                        "description": "",
                        "type": "integer",
                        "format": "int32",
                        "readOnly": false,
                        "nullable": true
                    },
                    "address": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "lat": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "lon": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "user_id": {
                        "description": "",
                        "type": "integer",
                        "format": "int32",
                        "readOnly": false,
                        "nullable": true
                    },
                    "duration": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "isConfirmed": {
                        "description": "",
                        "type": "boolean",
                        "readOnly": false,
                        "nullable": true
                    },
                    "hasAlreadySendRemeber": {
                        "description": "",
                        "type": "boolean",
                        "readOnly": false,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}